HUGE UPDATE:
Camera now follows the player and it is soft attached for areas in the beggining of the map. Do not leave outside map boundaries it could crash.
Fixed warnings in code as well.
Altered level rendering code so that it now renders the map properly on your screen at all times.